System-Tests: Add tests for cobbler-settings
Sadly the "simple" script we wrote was not so simple after all. Since the errors
we found were inside the script and not inside the main code, we need tests that
actually test the script itself. Since coverage for this script is irrelevant,
this is going into the integration tests.